+++ modules/rtos/src/GetEvent.c
@@ -136,7 +136,7 @@ StatusType GetEvent
 	{
 		SetError_Api(OSServiceId_SetEvent);
 		SetError_Param1(TaskID);
-		SetError_Param2(Event);
+		SetError_Param2((size_t)Event);
 		SetError_Ret(ret);
 		SetError_Msg("ActivateTask returns != than E_OK");
 		SetError_ErrorHook();
